front-knave/requirements.txt
fastapi
jinja2
httpx
python-dotenv
uvicorn
